// src/form/interface.tsx
export type IndexedObject = { [key: string]: any };
export type KeyType = string | number | symbol;
export type ComponentType = keyof JSX.IntrinsicElements | React.ComponentType<any>;

export type FieldError<FieldValue = any> = {
  value?: FieldValue;
  message?: ReactNode;
  type?: string;
  requiredError?: boolean;
};

export type ValidateFieldsErrors<FieldValue = any, FieldKey extends KeyType = string> =
  | Record<FieldKey, FieldValue>
  | undefined
  | null;

/**
 * @title Form
 */
export interface FormProps<
  FormData = any,
  FieldValue = FormData[keyof FormData],
  FieldKey extends KeyType = keyof FormData,
> extends Omit<FormHTMLAttributes<any>, 'className' | 'onChange' | 'onSubmit'> {
  style?: CSSProperties;
  className?: string | string[];
  prefixCls?: string;
  /**
   * @zh form
   * @en form
   */
  form?: FormInstance<FormData, FieldValue, FieldKey>;
  /**
   * @zh 设置后，会作为表单控件 `id`的前缀。
   * @en prefix of `id` attr
   */
  id?: string;
  /**
   * @zh 表单的布局，有三种布局，水平、垂直、多列。
   * @en The layout of Form
   * @defaultValue horizontal
   */
  layout?: 'horizontal' | 'vertical' | 'inline';
  /**
   * @zh 不同尺寸。
   * @en size of form
   */
  size?: 'mini' | 'small' | 'default' | 'large';
  /**
   * @zh
   * `<label>`标签布局，同[<Grid.Col>](/react/components/grid)组件接收的参数相同，
   * 可以配置`span`和`offset`值，会覆盖全局的`labelCol`设置
   * @en
   * Global `<label>` label layout. Same as the props received by the `<Grid.Col>`,
   * the values of `span` and `offset` can be configured,
   * which will be overwritten by the `labelCol` set by `Form.Item`
   * @defaultValue { span: 5, offset: 0 }
   */
  labelCol?: ColProps;
  /**
   * @zh
   * 控件布局，同`labelCol`的设置方法一致，会覆盖全局的`wrapperCol`设置，[ColProps](/react/components/grid)
   * @en
   * The global control layout, which is the same as the setting method of `labelCol`,
   * will be overwritten by the `wrapperCol` set by `Form.Item`
   * @defaultValue { span: 19, offset: 0 }
   */
  wrapperCol?: ColProps;
  /**
   * @zh 是否在 required 的时候显示加重的红色星号，设置 position 可选择将星号置于 label 前/后
   * @en Whether show red symbol when item is required，Set position props, you can choose to place the symbol before/after the label
   * @defaultValue true
   * @version `position` in 2.24.0
   */
  requiredSymbol?: boolean | { position: 'start' | 'end' };
  /**
   * @zh 标签的文本对齐方式
   * @en Text alignment of `label`
   * @defaultValue right
   */
  labelAlign?: 'left' | 'right';
  /**
   * @zh 设置表单初始值
   * @en Default value of form data
   */
  initialValues?: Partial<FormData>;
  /**
   * @zh
   * 触发验证的时机。
   * @en
   * When to trigger verification.
   * @defaultValue onChange
   * @version 2.28.0
   */
  validateTrigger?: string | string[];
  /**
   * @zh 任意表单项值改变时候触发。第一个参数是被改变表单项的值，第二个参数是所有的表单项值
   * @en Callback when any form item value changes.The first is the changed value, and the second is the value of all items
   */
  onValuesChange?: (value: Partial<FormData>, values: Partial<FormData>) => void;
  /**
   * @zh 表单项值改变时候触发。和 onValuesChange 不同的是只会在用户操作表单项时触发
   * @en Callback when the form item value changes. Unlike `onValuesChange`, it will only be called when the user manipulates the form item
   */
  onChange?: (value: Partial<FormData>, values: Partial<FormData>) => void;
  /**
   * @zh 配置最外层标签，可以是 html 标签或是组件
   * @en Custom outer tag. Can be html tags or React components
   * @defaultValue form
   */
  wrapper?: ComponentType;
  /**
   * @zh 配置 `wrapper` 之后，可以传一些参数到 wrapper 上。
   * @en If set `wrapper`, You can pass some parameters to the wrapper.
   */
  wrapperProps?: IndexedObject;
  /**
   * @zh 统一配置表单控件是否可用
   * @en Whether All Form item is disabled
   */
  disabled?: boolean;
  /**
   * @zh 是否显示标签后的一个冒号，优先级小于 `Form.Item` 中 `colon` 的优先级。(`ReactNode` in `v2.41.0`)
   * @en Whether show colon after `label`. Priority is lower than `colon` in `Form.Item`.(`ReactNode` in `v2.41.0`)
   */
  colon?: boolean | ReactNode;
  /**
   * @zh 验证失败后滚动到第一个错误字段。(`ScrollIntoViewOptions` 类型在 `2.19.0` 开始支持)
   * @en Whether scroll to first error item after validation fails. (`ScrollIntoViewOptions` is supported at `2.19.0`)
   */
  scrollToFirstError?: boolean | ScrollIntoViewOptions;
  /**
   * @zh 校验提示信息模板 [demo](/react/components/form#表单校验信息模板)
   * @en validation prompt template [demo](/react/en-US/components/form#validate%20messages)
   * @version 2.32.0
   */
  validateMessages?: Partial<{
    [key in keyof ValidateMessagesTemplateType]: ValidateMessagesTemplateType[key] extends string
      ? ValidateMessagesTemplateType[key] | ((data, { label }) => any)
      : Record<keyof ValidateMessagesTemplateType[key], string | ((data, { label }) => any)>;
  }>;
  /**
   * @zh 数据验证成功后回调事件
   * @en Callback when submit data
   */
  onSubmit?: (values: FormData) => void;
  /**
   * @zh 数据验证失败后回调事件
   * @en Callback when validate fail
   * @version 2.21.0
   */
  onSubmitFailed?: (errors: { [key: string]: FieldError }) => void;
}
// ...
export type FormInstance<
  FormData = any,
  FieldValue = FormData[keyof FormData],
  FieldKey extends KeyType = keyof FormData,
> = Pick<
  Store<FormData, FieldValue, FieldKey>,
  | 'getFieldsValue'
  | 'getFieldValue'
  | 'getFieldError'
  | 'getFieldsError'
  | 'getTouchedFields'
  | 'getFields'
  | 'setFieldValue'
  | 'setFieldsValue'
  | 'setFields'
  | 'resetFields'
  | 'clearFields'
  | 'submit'
  | 'validate'
> & {
  scrollToField: (field: FieldKey, options?: ScrollIntoViewOptions) => void;
  getInnerMethods: (inner?: boolean) => InnerMethodsReturnType<FormData, FieldValue, FieldKey>;
};

export type InnerMethodsReturnType<
  FormData = any,
  FieldValue = FormData[keyof FormData],
  FieldKey extends KeyType = keyof FormData,
> = Pick<
  Store<FormData, FieldValue, FieldKey>,
  | 'registerField'
  | 'registerWatcher'
  | 'innerSetInitialValues'
  | 'innerSetInitialValue'
  | 'innerSetCallbacks'
  | 'innerSetFieldValue'
  | 'innerGetStore'
>;